Create a separate homebrew for testing on higher firmware
For now I've just copied over everything I had on the normal one, tweaked it so that it passes on real hardware, and added an extra check for a libkernel error case that triggers on higher firmwares.
Also added a test for some behavior The Last of Us Part II does, since fpPS4 gets tripped up there for reasons I don't yet understand.