Use LONGS_EQUAL instead of CHECK_EQUAL
This means all fails now print in hex, which is significantly more convenient for just about every since thing here.
To make this work properly, I had to swap the error code defines to enums, as the easy to read hex representation of these error codes is too large for a signed integer.